Add files via upload
British Journal of Educational Technology Vol 34 No 3 2003 255–264
"Edutainment: is learning at risk?" - Zühal Okan

Educational technology defined
Technology is more than machinery, which maintains the existence and comfort of
humankind. Not a neutral tool, technology is loaded with cultural values:
“The new technology is not just an assemblage of machines and their accompanying software.
It embodies a form of thinking that orients a person to approach the world in a particular way.
Computers involve ways of thinking that under current educational conditions are primarily
technical. The more the new technology transforms the classroom into its own image, the more
a technical logic replaces critical, political and ethical understanding. The discourse of the classroom will centre on technique, and less on substance. Once again ‘how to’ will replace ‘why’.”
(Apple, 1991, 75)

As Apple notes, the debate about the impact of technology on education is usually
centred upon how it is used in the classroom, not taking the “why” into account. Thus,
the myth that technology is merely a tool at the teacher’s disposal is reinforced while
the ideological and ethical issues concerning what schools should be about and whose
interests they should serve are ignored.

Morrisett (1996) argues that society can be credited for creating technology, but technology is simultaneously creating society. People have become “compulsive information consumers” who favour the passive reception of information as a form of
entertainment over the more challenging act of thinking. Morrisett (1996) adds that
institutions of higher education have adapted to these conditions but, as a result, they
have also compromised the habits of mind (study, analysis, reflection, contemplation,
and deliberation) that are associated with logic."